requirements.dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.dev.in --universal --output-file ./requirements.dev.txt
anyio==4.8.0
    # via
    #   httpx
    #   jupyter-server
appnope==0.1.4 ; sys_platform == 'darwin'
    # via ipykernel
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==3.0.0
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via jupyterlab-server
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.2.0
    # via nbconvert
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   pyzmq
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via ipython
comm==0.2.2
    # via ipykernel
debugpy==1.8.12
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distlib==0.3.9
    # via virtualenv
entrypoints==0.4
    # via jupyter-client
et-xmlfile==2.0.0
    # via openpyxl
executing==2.2.0
    # via stack-data
fastjsonschema==2.21.1
    # via nbformat
filelock==3.18.0
    # via virtualenv
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via jupyterlab
identify==2.6.6
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==6.29.5
    # via jupyterlab
ipython==8.32.0
    # via ipykernel
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.10.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter-client==7.4.4
    # via
    #   -r ./requirements.dev.in
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.11.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.15.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.3.5
    # via -r ./requirements.dev.in
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via jupyterlab
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.1.1
    # via nbconvert
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via
    #   ipykernel
    #   jupyter-client
nodeenv==1.9.1
    # via pre-commit
notebook-shim==0.2.4
    # via jupyterlab
openpyxl==3.1.5
    # via -r ./requirements.dev.in
overrides==7.7.0
    # via jupyter-server
packaging==24.2
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via ipython
platformdirs==4.3.6
    # via
    #   jupyter-core
    #   virtualenv
pre-commit==4.1.0
    # via -r ./requirements.dev.in
prometheus-client==0.21.1
    # via jupyter-server
prompt-toolkit==3.0.50
    # via ipython
psutil==6.1.1
    # via ipykernel
ptyprocess==0.7.0 ; os_name != 'nt' or (sys_platform != 'emscripten' and sys_platform != 'win32')
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via
    #   ipython
    #   nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
python-json-logger==3.2.1
    # via jupyter-events
pywin32==308 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
    # via jupyter-core
pywinpty==2.0.14 ; os_name == 'nt'
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   terminado
pyyaml==6.0.2
    # via
    #   jupyter-events
    #   pre-commit
pyzmq==26.2.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruff==0.9.4
    # via -r ./requirements.dev.in
send2trash==1.8.3
    # via jupyter-server
setuptools==75.8.0
    # via jupyterlab
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.4.0
    # via bleach
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.12.2 ; python_full_version < '3.13'
    # via
    #   anyio
    #   referencing
uri-template==1.3.0
    # via jsonschema
urllib3==2.3.0
    # via requests
uv==0.5.26
    # via -r ./requirements.dev.in
virtualenv==20.29.1
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
